msm: kgsl: Place worker thread on SCHED_RR & WQ_HIGHPRI On commit 254852b (msm: kgsl: Increase worker thread priority), we had increased the priority for better rendering frames.
Since a lot of display related tasks are multiplexed on KGSL on a
dedicated thread, they are placed in a FIFO manner, which is a bit
of time consuming since not all those queued tasks have a known or
small execution period.

switching to Round-Robin fashion eliminates this complexity and bounds
these tasks on a time-slice and thus every tasks gets executed after
a certain time, as fast as possible.

Change-Id: I2b5137a5c6fdcc7fef7cafcc3e5728eab0034045
Signed-off-by: Panchajanya1999 <panchajanya@azure-dev.live>
Signed-off-by: Oktapra Amtono <oktapra.amtono@gmail.com>
Signed-off-by: negrroo <mohammedaelnaggar1@gmail.com>
Signed-off-by: onettboots <blackcocopet@gmail.com>